Fixed issues with pointnet examples
